src/functions/generate_reports/requirements.txt
functions-framework==3.7.0
google-cloud-firestore==2.17.2
google-cloud-storage==2.18.2
google-cloud-secret-manager==2.20.2
google-cloud-logging==3.10.0
pandas==2.2.2